Add Tween to position movement
I might end up taking this out later? I kind of like instant movement,
but I kinda like this too.

If it does end up getting taken out there should be a "and also, if
you're the player, update the camera" special clause built into this -
OR when we send a "moved" signal action whatever if that ever gets
hooked up.