Fire projectile action
So, hilariously, the projectiles continue PAST the player when
colliding. Because normally they'd explode when they touched the player,
but they don't because I didn't implement that, so they continue on the
path and the next square on the past is behind the player, so they
just...teleport past you.

Kinda fully, actually.