Add a Timeline class
Vaguely FFX-ish...

I'm deeply confused by my profiling results, because according to the
profiler results it should be a ~25% speedup (good but not enough, I
need to figure out wtf is going on) but I don't...notice it when it
runs.

Maybe I'm just bad at time. Or maybe I'm reading the profiler wrong?