Add non-functional level-up menu
I'd prefer it to be a modal, since this kind of raises awkward Rulebook
interactions and also it's just...janky, like the character select
screen will pop up when you kill a dude and it's confusing.

Modals seem unusually annoying in Godot though and I couldn't get the
Popup to work like I wanted so here we are.