Move SceneManager into singletons
The Godot integration with C# is kinda finnicky, the first time I moved
it I had edited the file in the editor but Godot apparently had its own
copy of the file? so when I moved it in Godot it dropped some code in
the file on the floor.

...'finnicky' to 'terrible' would describe a lot of the edge-casier
usages of Godot with C#. I mean, again, it doesn't have a way for you to
run tests! not great!