Add a background image
I'm not totally satisfied with how it looks or is implemented - for one
thing it stops when it reaches the border of the map! I need to figure
out how to handle when you run up against the edge of the map.

For the other thing it's a really hacky way of putting in a background
lol